+++ lib/weapons.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'adventure/weapon'
+
+require 'weapons/martial_melee1'
+require 'weapons/martial_melee2'
+require 'weapons/martial_ranged'
+require 'weapons/simple_melee'
+require 'weapons/simple_ranged'
+
+module Weapons
+  class Error < StandardError; end
+end